requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
requests==2.31.0
openai==1.109.1
httpx==0.28.1
pydantic==2.6.1
python-dotenv==1.0.1
